+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [2.0.5](https://github.com/SocketDev/socket-sdk-js/releases/tag/v2.0.5) - 2025-10-22
+
+### Added
+- `SocketSdkData<T>` type helper for extracting data from SDK operation results
+- `SocketSdkArrayElement<T, K>` type helper for extracting array element types from SDK operations
+
 ## [2.0.4](https://github.com/SocketDev/socket-sdk-js/releases/tag/v2.0.4) - 2025-10-22
 
 ### Added